+++ src/main.rs
@@ -15,10 +15,11 @@ mod utils;
 async fn main() -> Result<()> {
     match CompleteConfig::new() {
         Ok(config) => {
-            let app = App::new(config.terminal.maximum_messages as usize);
+            let app = App::new(config.terminal.maximum_messages);
 
             let (twitch_tx, terminal_rx) = mpsc::channel(100);
             let (terminal_tx, twitch_rx) = mpsc::channel(100);
+
             let cloned_config = config.clone();
 
             tokio::task::spawn(async move {
